+++ etna/datasets/tsdataset.py
@@ -220,8 +220,8 @@ class TSDataset:
         return future_ts
 
     @staticmethod
-    def _check_exog(df: pd.DataFrame, df_exog: pd.DataFrame):
-        """Check that df_exog have more timestamps than df."""
+    def _check_regressors(df: pd.DataFrame, df_exog: pd.DataFrame):
+        """Check that regressors in df_exog begin not later than in df and end later than in df."""
         df_segments = df.columns.get_level_values("segment")
         for segment in df_segments:
             target = df[segment]["target"].dropna()
@@ -242,8 +242,8 @@ class TSDataset:
                     )
 
     def _merge_exog(self, df: pd.DataFrame) -> pd.DataFrame:
-        self._check_exog(df=df, df_exog=self.df_exog)
-        df = pd.merge(df, self.df_exog, left_index=True, right_index=True).sort_index(axis=1, level=(0, 1))
+        self._check_regressors(df=df, df_exog=self.df_exog)
+        df = pd.merge(df, self.df_exog, left_index=True, right_index=True, how="left").sort_index(axis=1, level=(0, 1))
         return df
 
     def _check_endings(self):
@@ -313,7 +313,9 @@ class TSDataset:
         """
         return self._regressors
 
-    def plot(self, n_segments: int = 10, column: str = "target", segments: Optional[Sequence] = None, seed: int = 1):
+    def plot(
+        self, n_segments: int = 10, column: str = "target", segments: Optional[Sequence[str]] = None, seed: int = 1
+    ):
         """Plot of random or chosen segments.
 
         Parameters
@@ -341,6 +343,58 @@ class TSDataset:
             ax[i].set_title(segment)
 
         plt.show()
+
+    @staticmethod
+    def to_flatten(df: pd.DataFrame) -> pd.DataFrame:
+        """Return pandas DataFrame with flatten index.
+
+        Parameters
+        ----------
+        df:
+            DataFrame in ETNA format.
+
+        Returns
+        -------
+        pd.DataFrame
+            with TSDataset data
+
+        Examples
+        --------
+        >>> from etna.datasets import generate_const_df
+        >>> df = generate_const_df(
+        ...    periods=30, start_time="2021-06-01",
+        ...    n_segments=2, scale=1
+        ... )
+        >>> df.head(5)
+            timestamp    segment  target
+        0  2021-06-01  segment_0    1.00
+        1  2021-06-02  segment_0    1.00
+        2  2021-06-03  segment_0    1.00
+        3  2021-06-04  segment_0    1.00
+        4  2021-06-05  segment_0    1.00
+        >>> df_ts_format = TSDataset.to_dataset(df)
+        >>> TSDataset.to_flatten(df_ts_format).head(5)
+           timestamp  target    segment
+        0 2021-06-01     1.0  segment_0
+        1 2021-06-02     1.0  segment_0
+        2 2021-06-03     1.0  segment_0
+        3 2021-06-04     1.0  segment_0
+        4 2021-06-05     1.0  segment_0
+        """
+        aggregator_list = []
+        category = []
+        segments = df.columns.get_level_values("segment").unique().tolist()
+        for segment in segments:
+            if df[segment].select_dtypes(include=["category"]).columns.to_list():
+                category.extend(df[segment].select_dtypes(include=["category"]).columns.to_list())
+            aggregator_list.append(df[segment].copy())
+            aggregator_list[-1]["segment"] = segment
+        df = pd.concat(aggregator_list)
+        df = df.reset_index()
+        category = list(set(category))
+        df[category] = df[category].astype("category")
+        df.columns.name = None
+        return df
 
     def to_pandas(self, flatten: bool = False) -> pd.DataFrame:
         """Return pandas DataFrame.
@@ -373,12 +427,12 @@ class TSDataset:
         >>> df_ts_format = TSDataset.to_dataset(df)
         >>> ts = TSDataset(df_ts_format, "D")
         >>> ts.to_pandas(True).head(5)
-        feature  timestamp  target    segment
-        0       2021-06-01    1.00  segment_0
-        1       2021-06-02    1.00  segment_0
-        2       2021-06-03    1.00  segment_0
-        3       2021-06-04    1.00  segment_0
-        4       2021-06-05    1.00  segment_0
+           timestamp  target    segment
+        0 2021-06-01     1.0  segment_0
+        1 2021-06-02     1.0  segment_0
+        2 2021-06-03     1.0  segment_0
+        3 2021-06-04     1.0  segment_0
+        4 2021-06-05     1.0  segment_0
         >>> ts.to_pandas(False).head(5)
         segment    segment_0 segment_1
         feature       target    target
@@ -391,19 +445,7 @@ class TSDataset:
         """
         if not flatten:
             return self.df.copy()
-        if flatten:
-            aggregator_list = []
-            category = []
-            for segment in self.segments:
-                if self.df[segment].select_dtypes(include=["category"]).columns.to_list():
-                    category.extend(self.df[segment].select_dtypes(include=["category"]).columns.to_list())
-                aggregator_list.append(self.df[segment].copy())
-                aggregator_list[-1]["segment"] = segment
-            df = pd.concat(aggregator_list)
-            df = df.reset_index()
-            category = list(set(category))
-            df[category] = df[category].astype("category")
-            return df
+        return self.to_flatten(self.df)
 
     @staticmethod
     def to_dataset(df: pd.DataFrame) -> pd.DataFrame:
@@ -679,7 +721,7 @@ class TSDataset:
         """
         return self.df.head(n_rows)
 
-    def tail(self, n_rows: Optional[int] = None) -> pd.DataFrame:
+    def tail(self, n_rows: int = 5) -> pd.DataFrame:
         """Return the last `n` rows.
 
         Mimics pandas method.