+++ etna/transforms/lags.py
@@ -11,7 +11,7 @@ from etna.transforms.base import Transform
 class _OneSegmentLagFeature(Transform):
     """Generates series of lags from given segment."""
 
-    def __init__(self, lags: Union[List[int], int], in_column: str, out_column: Optional[str] = None):
+    def __init__(self, in_column: str, lags: Union[List[int], int], out_column: Optional[str] = None):
         if isinstance(lags, int):
             if lags < 1:
                 raise ValueError(f"{type(self).__name__} works only with positive lags values, {lags} given")
@@ -37,15 +37,15 @@ class _OneSegmentLagFeature(Transform):
 class LagTransform(PerSegmentWrapper):
     """Generates series of lags from given dataframe."""
 
-    def __init__(self, lags: Union[List[int], int], in_column: str, out_column: Optional[str] = None):
+    def __init__(self, in_column: str, lags: Union[List[int], int], out_column: Optional[str] = None):
         """Create instance of LagTransform.
 
         Parameters
         ----------
-        lags:
-            int value or list of values for lags computation; if int, generate range of lags from 1 to given value
         in_column:
             name of processed column
+        lags:
+            int value or list of values for lags computation; if int, generate range of lags from 1 to given value
         out_column:
             name of added column. We get '{out_column}_{lag_number}'(don't forget to add regressor prefix if necessary).
             If not given, use 'regressor_{self.__repr__()}_{lag_number}'
@@ -55,14 +55,14 @@ class LagTransform(PerSegmentWrapper):
         ValueError:
             if lags value contains non-positive values
         """
-        self.lags = lags
         self.in_column = in_column
+        self.lags = lags
         self.out_column = out_column
 
         super().__init__(
             transform=_OneSegmentLagFeature(
-                lags=self.lags,
                 in_column=self.in_column,
+                lags=self.lags,
                 out_column=out_column if out_column is not None else f"regressor_{self.__repr__()}",
             )
         )