+++ tests/test_models/test_sarimax_model.py
@@ -37,9 +37,9 @@ def test_sarimax_forecaster_run_with_reg(example_reg_tsds):
 
 
 def test_confidence_interval_run_insample(example_tsds):
-    model = SARIMAXModel(interval_width=0.95)
+    model = SARIMAXModel()
     model.fit(example_tsds)
-    forecast = model.forecast(example_tsds, confidence_interval=True)
+    forecast = model.forecast(example_tsds, confidence_interval=True, interval_width=0.95)
     for segment in forecast.segments:
         segment_slice = forecast[:, segment, :][segment]
         assert {"target_lower", "target_upper", "target"}.issubset(segment_slice.columns)
@@ -47,10 +47,10 @@ def test_confidence_interval_run_insample(example_tsds):
 
 
 def test_confidence_interval_run_infuture(example_tsds):
-    model = SARIMAXModel(interval_width=0.95)
+    model = SARIMAXModel()
     model.fit(example_tsds)
     future = example_tsds.make_future(10)
-    forecast = model.forecast(future, confidence_interval=True)
+    forecast = model.forecast(future, confidence_interval=True, interval_width=0.95)
     for segment in forecast.segments:
         segment_slice = forecast[:, segment, :][segment]
         assert {"target_lower", "target_upper", "target"}.issubset(segment_slice.columns)